python/requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-mail==1.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
jose==1.0.0
MarkupSafe==2.1.5
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
pyodbc==5.1.0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
